Fix partition values to not suck

+++ util/sieve/highest_prime.h
@@ -2,7 +2,7 @@ using namespace std;
 
 class HighestPrimeSieve {
 
-    vector<uint64_t> primes;
+    vector<uint32_t> primes;
 
 public:
 
@@ -11,7 +11,7 @@ public:
         init_sieve();
     }
 
-    uint64_t count() {
+    uint32_t count() {
         int count = 0;
         for (int i = 1; i < primes.size(); i++) {
             count += primes[i] == 0;
@@ -20,23 +20,23 @@ public:
     }
 
     void print() {
-        for (uint64_t i = 0; i < primes.size(); i++) {
+        for (uint32_t i = 0; i < primes.size(); i++) {
             cout << i << " " << primes[i] << endl;
         }
     }
 
-    uint64_t highest_prime(uint64_t n) {
+    uint32_t highest_prime(uint32_t n) {
         return primes[n];
     }
 
-    bool is_prime(uint64_t n) {
+    bool is_prime(uint32_t n) {
         return primes[n] == 0;
     }
 
     // Returns the next prime greater than n
     // returns 0 if next prime isn't within sieve.
-    int next_prime(uint64_t n) {
-        for (uint64_t i = n + 1; i < primes.size(); i++) {
+    int next_prime(uint32_t n) {
+        for (uint32_t i = n + 1; i < primes.size(); i++) {
             if (primes[i] == 0) { return i; }
         }
         return 0;
@@ -45,9 +45,9 @@ public:
 private:
 
     void init_sieve() {
-        for (uint64_t i = 2; i < primes.size(); i++) {
+        for (uint32_t i = 2; i < primes.size(); i++) {
             if (primes[i] == 0) {
-                for (uint64_t j = 2 * i; j < primes.size(); j += i) {
+                for (uint32_t j = 2 * i; j < primes.size(); j += i) {
                     primes[j] = i;
                 }
             }
